requirements.txt
fastapi==0.116.2
numpy==2.3.3
pandas==2.3.2
Pillow==11.3.0
pydantic==2.11.9
Requests==2.32.5
scikit_learn==1.7.2
streamlit==1.49.1
torch==2.8.0
torchvision==0.23.0
uvicorn==0.36.0
tensorboard==2.20.0
